Remove trait or trait combination from possible trait combinations of possible Trait combinations

Remove trait or trait combination from possible trait combinations of possible Trait combinations

Usage

rcomb.indices(ncomb, i, pmat, gmat, wmat, wcol = 1, GAY)

Arguments

ncomb

Number of character combination

i

remove trait or trait combination

pmat

Phenotypic Variance Covariance Matrix

gmat

Genotypic Variance Covariance Matrix

wmat

Weight Matrix

wcol

Respective weight column number of Weight Matrix

GAY

Genetic Advance/Genetic Gain of base selection index

Value

Data frame of possible selection indices with per cent relative efficiency and ranking

Examples

gmat<- gen.varcov(seldata[,3:9], seldata[,2], seldata[,1])
pmat<- phen.varcov(seldata[,3:9], seldata[,2], seldata[,1])
rcomb.indices(ncomb = 2, i = 1, pmat = pmat, gmat = gmat, wmat = weight[,2:3], wcol = 1)
#>      ID    b.1     b.2     GA PRE Rank
#> 1  2, 3 0.4139  1.1055 2.1197 100    8
#> 2  2, 4 0.3435  0.1654 1.3321 100    8
#> 3  2, 5 0.3718  0.2117 1.6600 100    8
#> 4  2, 6 0.4170  1.3135 1.9304 100    8
#> 5  2, 7 0.4266  2.2783 1.8200 100    8
#> 6  3, 4 0.5322  0.4984 0.8230 100    8
#> 7  3, 5 1.7692  0.1081 1.9993 100    8
#> 8  3, 6 0.5428  0.0429 0.5365 100    8
#> 9  3, 7 0.5277 -0.0789 0.5200 100    8
#> 10 4, 5 0.9933  0.2046 1.8449 100    8
#> 11 4, 6 0.4785  0.0755 0.6720 100    8
#> 12 4, 7 0.5270  0.6339 0.7809 100    8
#> 13 5, 6 0.2208 -0.7865 1.0055 100    8
#> 14 5, 7 0.2008 -1.9018 1.0765 100    8
#> 15 6, 7 0.0808  0.7393 0.2609 100    8