selection.index: Analysis of Selection Index in Plant Breeding

Provides tools for the simultaneous improvement of multiple traits in plant breeding. Building upon the classical selection index (Smith 1937 doi:10.1111/j.1469-1809.1936.tb02143.x ) and modern quantitative genetics (Kang 2020 doi:10.1007/978-3-319-91223-3 ), this package calculates classical phenotypic, genomic, marker-assisted, restricted/constrained, and eigen selection indices. It also incorporates multi-stage selection evaluation and stochastic simulations to estimate genetic advance based on economic weights, heritability, and genetic correlations.

See also

Author

Maintainer: Zankrut Goyani zankrut20@gmail.com [copyright holder]